+++ Engine/src/Engine/GameObject/ScriptableBase.hpp
@@ -7,10 +7,9 @@ namespace Engine {
     public:
         virtual ~ScriptableBase() {}
         
+        // Helper functions for all native scripts
         template<typename Component>
-        Component& GetComponent() {
-            return m_GameObject.GetComponent<Component>();
-        }
+        Component& GetComponent() { return m_GameObject.GetComponent<Component>(); }
 
         GameObject_type GetGameObjectID() const { return m_GameObject.GetHandle(); }
